requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
mem0ai==1.0.0
python-dotenv==1.0.1
pydantic==2.9.2
openai>=1.90.0
qdrant-client==1.12.1
redis==5.2.0
boto3==1.35.36
python-multipart==0.0.12
requests==2.31.0

